python hindi iter_1

+++ analyze/ClassEval_59::add_movie.py
@@ -15,6 +15,6 @@ def add_movie(self, name, price, start_time, end_time, n):
         """
     start_dt = datetime.strptime(start_time, '%H:%M')
     end_dt = datetime.strptime(end_time, '%H:%M')
-    seats = np.zeros((n, n), dtype=float)
+    seats = np.zeros((n, n))
     movie = {'name': name, 'price': float(price), 'start_time': start_dt, 'end_time': end_dt, 'seats': seats}
     self.movies.append(movie)
